+++ src/main/java/com/github/j3t/mvnio/maven/validation/VersionValidator.java
@@ -1,10 +1,10 @@
-package com.github.j3t.mvnio.repo.validation;
+package com.github.j3t.mvnio.maven.validation;
 
 import reactor.core.publisher.Mono;
 
 import java.util.regex.Pattern;
 
-import static com.github.j3t.mvnio.repo.validation.IdValidator.ID;
+import static com.github.j3t.mvnio.maven.validation.IdValidator.ID;
 import static java.lang.String.format;
 
 /**
